Restructure huginn-net-db (#273)
Restructure the database in order to simplify features
