+++ huginn-net-db/src/tcp/matching.rs
@@ -1,10 +1,29 @@
-use crate::db::TcpIndexKey;
+//! Glue between [`huginn_net_tcp::observable::TcpObservation`] and the
+//! database matcher infrastructure.
+//!
+//! Provides:
+//! - `pub(crate)` distance helpers that read fields off a `TcpObservation`
+//!   (`distance_olen`, `distance_mss`, `distance_wscale`, `distance_olayout`,
+//!   `distance_quirks`). The pure helpers that compare two raw signature
+//!   types ([`crate::tcp::distance_ttl`], [`crate::tcp::distance_window_size`],
+//!   …) live in `crate::tcp::distances` (private module; re-exported through
+//!   [`crate::tcp`]).
+//! - The [`crate::db_matching_trait::ObservedFingerprint`] impl that turns an
+//!   observation into a [`crate::database::TcpIndexKey`].
+//! - The [`crate::db_matching_trait::DatabaseSignature`] impl that scores a
+//!   `tcp::Signature` against a `TcpObservation`.
+//!
+//! For backward compatibility this module is re-exposed at the crate root as
+//! `huginn_net_db::observable_tcp_signals_matching` via a `#[path]` shim in
+//! `lib.rs`.
+
+use crate::database::TcpIndexKey;
 use crate::db_matching_trait::{DatabaseSignature, MatchQuality, ObservedFingerprint};
-use crate::observable_signals::TcpObservation;
 use crate::tcp::{
     self, distance_ip_version, distance_payload_size, distance_ttl, distance_window_size,
     IpVersion, PayloadSize,
 };
+use huginn_net_tcp::observable::TcpObservation;
 
 pub(crate) fn distance_olen(observed: &TcpObservation, signature: &tcp::Signature) -> Option<u32> {
     if observed.olen == signature.olen {
